Rename YamlExcludedLabelsRepository to InMemoryExcludedLabelsRepository
